requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
tqdm>=4.62.0
biopython>=1.79
requests>=2.26.0
plotly>=5.3.0
jinja2>=3.0.0 